Add Products component
The component that will show the products was added, it receives the
products array and returns a list with all the products in the array.

The service folder was created to hold the products.js which have a
function that fetches all the products available in the dummyJSON API
and returns them as an array.

Also the Icons component was created to hold all the svg icons that will
be used in the project.