Consume the filters context
With the previous creation of the filters context we only need for the
components to consume it, to this we first need to put the children
components inside the context PROVIDER, with this all the components
inside this provider will have access to the value we defined for the
context.

This context value is first used by the useFilters custom hook, and then
this hook is implemented where is needed (Filters component and the
App.jsx).

I understood this implementation of useContext like this: as context
provider have the filters state, every time this state change, the
components using it will be rendered again, so when the user change the
starting price or the category for the products the App will be rendered
again and specifically will use the filterProducts function passed by
the useFilters hook, and then will provide the products that have meet
those filters changes. In a nutshell when the filters state from the
context provider changes, will cause re-renders in the components that
have access to that context, in this case the entire App.jsx.