Create the cart item component
The react component for the items of the cart was created and the
corresponding styles were added. This currently does not work and is
a temporal way to visualize how the cart items will be displayed.

The Icons component was modified, specifically its AddToCartIcon, to now
return a more accurate svg.